simple_authorization_server/requirements.txt
fastapi==0.78.0
uvicorn[standard]==0.17.6
python-multipart==0.0.5
google-cloud-iam==0.3.1
